mini_blog/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
pydantic==2.11.7
python-dotenv==1.1.1
